separate multiprocess test
